+++ changelog.md
@@ -51,7 +51,7 @@
 
 ### Changed
 
-- Overhauled Access Transformer support:
+- Overhauled Access Transformer and Access Widener support:
   - many lexing errors should now be fixed
   - class names and member names now have their own references, replacing the custom Goto handler
   - SRG names are no longer used on NeoForge 1.20.2+ and a new copy action is available for it